cli: add and test `debug decode-value`
Inspired by #28995, where I needed this.

Release note: None
